% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{backsolveSNPeff}
\alias{backsolveSNPeff}
\title{backsolveSNPeff}
\usage{
backsolveSNPeff(Z, g)
}
\arguments{
\item{Z}{Centered marker matrix (dominance deviations must also be centered)}

\item{g}{The solutions (blups, i.e. GEBVs) from the GBLUP model}
}
\value{
matrix of SNP effects matching RR-BLUP / SNP-BLUP
}
\description{
From the GBLUP solutions and a centered SNP matrix backsolve SNP effects
}
\examples{
A<-kinship(M,type="add")
trainingDF \%<>\% dplyr::mutate(ga=factor(as.character(id),
                                        levels=rownames(A)),
                              gd=ga)
gblup<-mmer(pheno~1,
            random=~vs(ga,Gu = A),
            weights=WT,
            data=trainingDF,verbose = T)
ga<-as.matrix(gblup$U$`u:ga`$pheno,ncol=1)
Za<-centerDosage(M)
snpeff<-backsolveSNPeff(Za,ga)
}
