% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{centerDosage}
\alias{centerDosage}
\title{centerDosage}
\usage{
centerDosage(M)
}
\arguments{
\item{M}{dosage matrix. Assumes SNPs in M coded 0, 1, 2 (requires rounding dosages to integers). M is Nind x Mrow, numeric matrix, with row/columanes to indicate SNP/ind ID}
}
\value{

}
\description{
Centers dosage matrix, e.g. for use in whole-genome regressions like rrBLUP
}
\examples{
centeredM<-centerDosage(M)
}
