% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{crosses2predict}
\alias{crosses2predict}
\title{crosses2predict}
\usage{
crosses2predict(parents)
}
\arguments{
\item{parents}{}
}
\value{
tibble, two columns, sireID and damID, all pairwise crosses (see details).
}
\description{
Make a data.frame of all pairwise matings given a vector of parent IDs.
Include selfs. No reciprocal crosses, i.e. use as male == use as female.
Diagonal and upper-triangle of mating matrix.
}
