% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dose2domDev}
\alias{dose2domDev}
\title{dose2domDev}
\usage{
dose2domDev(M)
}
\arguments{
\item{M}{dosage matrix. Assumes SNPs in M coded 0, 1, 2 (requires rounding dosages to integers). M is Nind x Mrow, numeric matrix, with row/columanes to indicate SNP/ind ID}
}
\value{

}
\description{
Converts a dosage matrix into a matrix of centered dominance deviations
}
\examples{
domDev<-dose2domDev(M)
}
