% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dose2domDevGenotypic}
\alias{dose2domDevGenotypic}
\title{dose2domDevGenotypic}
\usage{
dose2domDevGenotypic(M)
}
\arguments{
\item{M}{dosage matrix. Assumes SNPs in M coded 0, 1, 2 (requires rounding dosages to integers). M is Nind x Mrow, numeric matrix, with row/columanes to indicate SNP/ind ID}
}
\value{

}
\description{
Converts a dosage matrix into a matrix of centered dominance deviations.
This sets up the "genotypic" (aka "biological") partition of additive dominance in terms of their genotypic effects instead of in terms of breeding values or dominance deviations.
See function dose2domDev() to set-up the "statistical" (aka "classical") partition in terms of genotypic effects.
See Vitezica et al. 2013.
Also see Varona et al. 2018. https://doi.org/10.3389/fgene.2018.00078
}
\examples{
domDev<-dose2domDevGenotypic(M)
}
