% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{effectsArray2list}
\alias{effectsArray2list}
\title{effectsArray2list}
\usage{
effectsArray2list(effectsArray, snpIDs, traits, nIter, burnIn, thin)
}
\arguments{
\item{effectsArray}{According to BGLR documentation: 3D array, with dim=c(nRow,p,traits), where nRow number of MCMC samples saved, p is the number of predictors and traits is the number of traits. BGLR::Multitrait() writes a binary file to disk when saveEffects=TRUE is specified. It can be read to R with BGLR::readBinMatMultitrait().}

\item{snpIDs}{character vector with labels for the predictors (SNPs), numeric should work too, but untested.}

\item{traits}{character vector to label the traits.}

\item{nIter}{number of iterations used for MCMC; used internally only to exclude burn-in samples from computation}

\item{burnIn}{burnIn for MCMC; used internally only to exclude burn-in samples from computation}

\item{thin}{thin for MCMC; used internally only to exclude burn-in samples from computation}
}
\value{
list of matrices, one matrix per trait, each matrix has `nrow((nIter-burnIn)/thin)` and `ncol(length(snpIDs))`. Each element of the list is named with a string identifying the trait and the colnames of each matrix are labelled with snpIDs.
}
\description{
Converts an array of posterior samples of multi-trait marker effects to a named list (one for each trait).
}
\examples{
effectsArray<-BGLR::readBinMatMultitrait(filename = "mt_effects_ETA_g_beta.bin")
effectsArray<-effectsArray2list(effectsArray, snpIDs, traits, nIter, burnIn, thin)

}
