% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{genmap2recombfreq}
\alias{genmap2recombfreq}
\title{genmap2recombfreq}
\usage{
genmap2recombfreq(m)
}
\arguments{
\item{m}{vector of centiMorgan-scale genetic positions. names(m) correspond to a SNP_ID. Since m potentially contains all chromosomes, sets recomb. freq. b/t chrom. to 0.5}
}
\value{
potentially really large matrix of pairwise recombination frequencies between loci
}
\description{
Compute the pairwise recombination frequencies between all loci from genetic map positions.
}
\details{
May be worth computing in an R session using multi-threaded BLAS
}
