% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{genomicVarD_m1}
\alias{genomicVarD_m1}
\title{genomicVarD_m1}
\usage{
genomicVarD_m1(snpVarD, freq)
}
\arguments{
\item{snpVarD}{vector of _dominance_ SNP effects estimate. Ideally vector is named with SNP IDs and matches the order of related SNP matrices.}

\item{freq}{vector of allele frequencies. Must be in same order as snp effects}
}
\value{
additive genomic variance component
}
\description{
Compute standard genomic variance term, what Lehermeier called Method 1 or "M1".
LD is not accounted for. SNP effects are assumed _i.i.d._.
}
