% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{genoVarCovarMatFunc}
\alias{genoVarCovarMatFunc}
\title{genoVarCovarMatFunc}
\usage{
genoVarCovarMatFunc(Z)
}
\arguments{
\item{Z}{column-centered matrix of SNP dosages. Assumes SNPs in Z were originally coded 0, 1, 2 were column centered.}
}
\value{
NOTE: this matrix is going to be big in practice.
 The *p SNPs*-by-*p SNPs* variance-covariance matrix of SNP dosages.
 may be worth computing in an R session using multi-threaded BLAS
}
\description{
Compute the *p SNPs*-by-*p SNPs* variance-covariance matrix of SNP dosages.
This is an estimator of the LD between loci within a given population.
}
\examples{
Z<-centerDosage(M)
genoVarCovarMat<-genoVarCovarMatFunc(Z)
}
