% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar_frequentist_univar.R
\name{getM2varcomp}
\alias{getM2varcomp}
\title{getM2varcomp}
\usage{
getM2varcomp(effects, varcovarmat, type)
}
\arguments{
\item{effects}{vector of SNP effects}

\item{varcovarmat}{square, symmetric var-covar matrix, estimator of LD between loci}

\item{type}{string, "add" or "dom". If "dom" will square varcovarmat to get correct varcomp.}
}
\value{
additive or dominance genomic variance estimate accounting account for linkage disequilibrium
}
\description{
Computes the genomic variance component accounting account for linkage disequilibrium, which Lehermeier called Method 2 or "M2".
}
\details{
Make sure **effects** and **varcovarmat** are in same order.
This might use _alot_ of RAM and take _alot_ of time as it involves very large matrix operations.
May be worth computing in an R session using multi-threaded BLAS
}
\examples{
Z<-centerDosage(M)
varcovarmat<-genoVarCovarMatFunc(Z)
VarAddM2<-getM2varcomp(effects,varcovarmat,"add")
}
