% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{getMAF}
\alias{getMAF}
\title{getMAF}
\usage{
getMAF(M)
}
\arguments{
\item{M}{dosage matrix. Assumes SNPs in M coded 0, 1, 2 (requires rounding dosages to integers). M is Nind x Mrow, numeric matrix, with row/columanes to indicate SNP/ind ID}
}
\value{
vector of _minor_ allele frequencies, names = SNP IDs if in cols of M
}
\description{
get a vector of _minor_ allele frequences from a dosage matrix
}
