% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{getMultiTraitPMVs_AD}
\alias{getMultiTraitPMVs_AD}
\title{getMultiTraitPMVs_AD}
\usage{
getMultiTraitPMVs_AD(AddEffectList, DomEffectList, genoVarCovarMat)
}
\arguments{
\item{AddEffectList}{list of ADDITIVE effect matrices, one matrix per trait, Each element of the list is named with a string identifying the trait and the colnames of each matrix are labelled with snpIDs.}

\item{DomEffectList}{list of DOMINANCE effect matrices, one matrix per trait, Each element of the list is named with a string identifying the trait and the colnames of each matrix are labelled with snpIDs.}

\item{genoVarCovarMat}{variance-covariance matrix of marker genotypes, i.e. measuring linkage disequilibrium}
}
\value{
tibble with estimators for all pairwise variance and co-variance parameters including both additive and dominance components.
}
\description{
For an additive plus dominance model, compute _all pairwise_ variance and co-variance parameters from the post-burn-In, thinned samples of marker effects obtained using an Bayesian multi-trait model fit to some current population.
Wraps `posteriorMeanVarCovarAD()` over all trait variances and co-variances.
Posterior mean variance or co-variance (PMV) and also the variance of posterior means (VPM) estimators are returned.
Both M1, i.e. "Method 1", which does not account for LD in the population and M2 or "Method 2", which does, are returned.
Additive and dominance variances/covariances are returned.
}
