% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getPropHom}
\alias{getPropHom}
\title{getPropHom}
\usage{
getPropHom(M)
}
\arguments{
\item{M}{dosage matrix. Assumes SNPs in M coded 0, 1, 2 (requires rounding dosages to integers). M is Nind x Mrow, numeric matrix, with row/columanes to indicate SNP/ind ID}
}
\value{
vector of per-individual proportion homozygous
}
\description{
Compute the per-individual proportion homozygous.
For example, to use as a predictor for a directional dominance model.
}
