% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{intensity}
\alias{intensity}
\title{intensity}
\usage{
intensity(propSel)
}
\arguments{
\item{propSel}{proportion selection}
}
\value{

}
\description{
Compute the standardized selection intensity from the proportion selection.
}
