% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{kinship}
\alias{kinship}
\title{kinship function}
\usage{
kinship(M, type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Function to create additive and dominance genomic relationship matrices from biallelic dosages.
}
\examples{
K<-kinship(M,"add")
}
