% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{maf_filter}
\alias{maf_filter}
\title{maf_filter}
\usage{
maf_filter(M, thresh)
}
\arguments{
\item{M}{dosage matrix. Assumes SNPs in M coded 0, 1, 2 (requires rounding dosages to integers). M is Nind x Mrow, numeric matrix, with row/columanes to indicate SNP/ind ID}

\item{thresh}{threshold value. Columns of M with maf<thresh will be removed}
}
\value{
dosage matrix potentially with columns removed
}
\description{
filter a dosage matrix by minor allele frequence.
get a vector of allele frequences from a dosage matrix
}
