% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{makeGRM}
\alias{makeGRM}
\title{makeGRM function}
\usage{
makeGRM(pop, SP, type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
function to create a additive or dominance kinship matrix from a AlphaSimR pop-class object. A wrapper for the kinship() function.
}
\examples{
 K<-makeGRM(pop, SP, type="add")

}
