% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{makeKinship}
\alias{makeKinship}
\title{makeKinship function}
\usage{
makeKinship(pop, SP, type, snpsORqtl = "snps")
}
\arguments{
\item{pop}{pop-class object from AlphaSimR}

\item{SP}{simulation parameters from AlphaSimR}

\item{type}{string, "add" or "dom". type="add" gives same as rrBLUP::A.mat(), i.e. Van Raden, Method 1. type="dom" gives classical parameterization according to Vitezica et al. 2013.}

\item{snpsORqtl}{string, "snps" (default) or "qtl". Determines which loci are used to compute the GRM.}
}
\value{

}
\description{
function to create a additive or dominance kinship matrix from a AlphaSimR pop-class object. A wrapper for the kinship() function.
}
\examples{
 K<-makeKinship(pop, SP, type="add")

}
