% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{multitraitPMV_A}
\alias{multitraitPMV_A}
\title{multitraitPMV_AD}
\usage{
multitraitPMV_A(AddEffectArray, traits, genoVarCovarMat, nIter, burnIn, thin)
}
\arguments{
\item{AddEffectArray}{3D array of ADDITIVE marker effects from each MCMC sample. BGLR::Multitrait() writes a binary file to disk when saveEffects=TRUE is specified. It can be read to R with BGLR::readBinMatMultitrait().}

\item{traits}{character vector, in order of traits supplied to BGLR}

\item{genoVarCovarMat}{variance-covariance matrix of marker genotypes, i.e. linkage disequilibrium}

\item{nIter}{number of iterations used for MCMC [used internally only to exclude burn-in samples from computation]}

\item{burnIn}{burnIn for MCMC [used internally only to exclude burn-in samples from computation]}

\item{thin}{thin for MCMC [used internally only to exclude burn-in samples from computation]}
}
\value{
tibble with both M1 (no LD) and M2 (LD) for all variances and covariances
}
\description{
Compuate all pairwise _posterior mean_ genetic variances and covariance for an Bayesian multi-trait single kernel, additive effects-only model fit to some current population.
}
\details{
Lehermeier et al. 2017. J Anim Breed Genet. 2017;134:232–41.
Lehermeier et al. 2017. Genetics, 207(4), 1651–1661. https://doi.org/10.1534/genetics.117.300403
Neyhart et al. 2019. G3 9(10), https://doi.org/10.1534/g3.119.400406
Alves et al. 2019. Plant Methods, https://doi.org/10.1186/s13007-019-0388-x
Schreck et al. 2019. https://doi.org/10.1534/genetics.119.302324
}
