% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{predCrossMeanGVsOneTrait}
\alias{predCrossMeanGVsOneTrait}
\title{predCrossMeanGVsOneTrait}
\usage{
predCrossMeanGVsOneTrait(
  Trait,
  CrossesToPredict,
  doseMat,
  postMeanAddEffects,
  postMeanDomEffects
)
}
\arguments{
\item{Trait}{string, label of trait (name in list of postMeanAddEffects) to compute}

\item{CrossesToPredict}{data.frame or tibble, col/colnames: sireID, damID. sireID and damID must both be in the haploMat.}

\item{doseMat}{dosage matrix. Assumes SNPs in M coded 0, 1, 2 (requires rounding dosages to integers). M is Nind x Mrow, numeric matrix, with row/colnames to indicate SNP/ind ID}

\item{postMeanAddEffects}{list of named vectors (or column matrices) with the posterior mean ADDITIVE marker effects.}

\item{postMeanDomEffects}{list of named vectors (or column matrices) with the posterior mean DOMINANCE marker effects.}
}
\value{
tibble with predicted mean GV for each family
}
\description{
Predict the total genetic merit of the cross based on a model with additive+dominance marker effects.
For each family , for a single trait, given  parental allelic dosages and (posterior mean) marker effects.
G = sum( pr(AA)*a-pr(aa)*a+pr(Aa)*d ), where A is counted allele in dosages
}
