% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{predCrossMeansA}
\alias{predCrossMeansA}
\title{predCrossMeansA}
\usage{
predCrossMeansA(CrossesToPredict, postMeanAddEffects, doseMat)
}
\arguments{
\item{CrossesToPredict}{data.frame or tibble, col/colnames: sireID, damID. sireID and damID must both be in the haploMat.}

\item{postMeanAddEffects}{list of named vectors (or column matrices) with the posterior mean ADDITIVE marker effects.}

\item{doseMat}{dosage matrix. Assumes SNPs in M coded 0, 1, 2 (requires rounding dosages to integers). M is Nind x Mrow, numeric matrix, with row/colnames to indicate SNP/ind ID}
}
\value{
tibble with predicted mean BV for each trait in each family
}
\description{
From an additive only model, fit to multiple traits, predict the mean (breeding value) of each cross for each trait.
Corresponds to the mean GEBV of the parents, given  parental allelic dosages and (posterior mean) marker effects.
}
