% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{predCrossMeansAD}
\alias{predCrossMeansAD}
\title{predCrossMeansAD}
\usage{
predCrossMeansAD(
  CrossesToPredict,
  postMeanAddEffects,
  postMeanDomEffects,
  doseMat
)
}
\arguments{
\item{CrossesToPredict}{data.frame or tibble, col/colnames: sireID, damID. sireID and damID must both be in the haploMat.}

\item{postMeanAddEffects}{list of named vectors (or column matrices) with the posterior mean ADDITIVE marker effects.}

\item{postMeanDomEffects}{list of named vectors (or column matrices) with the posterior mean DOMINANCE marker effects.}

\item{doseMat}{dosage matrix. Assumes SNPs in M coded 0, 1, 2 (requires rounding dosages to integers). M is Nind x Mrow, numeric matrix, with row/colnames to indicate SNP/ind ID}
}
\value{
tibble with predicted mean GV for each trait in each family
}
\description{
From an additive+dominance model, fit to multiple traits, predict the total genetic merit of each cross for each trait.
For each family , for a single trait, given  parental allelic dosages and (posterior mean) marker effects.
G = sum( ğ‘(ğ‘ âˆ’ ğ‘ âˆ’ ğ‘¦) + ğ‘‘[2ğ‘ğ‘ + ğ‘¦(ğ‘ âˆ’ ğ‘)] )
a and d being the additive and dominance effects
p and q being the allele frequencies of one parent
y is the difference of freq. between the two parents
}
