% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{remove_invariant}
\alias{remove_invariant}
\title{remove_invariant}
\usage{
remove_invariant(M)
}
\arguments{
\item{M}{dosage matrix. Assumes SNPs in M coded 0, 1, 2 (requires rounding dosages to integers). M is Nind x Mrow, numeric matrix, with row/columanes to indicate SNP/ind ID}

\item{thresh}{threshold value. Columns of M with maf<thresh will be removed}
}
\value{
dosage matrix potentially with columns removed
}
\description{
filter a dosage matrix, removing invariant markers. Removes e.g. cases where MAF=0.5 but all dosages == 1 (het).
}
