% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar_frequentist_univar.R
\name{runCrossVarPredsAD}
\alias{runCrossVarPredsAD}
\title{runCrossVarPredsAD}
\usage{
runCrossVarPredsAD(
  outprefix = NULL,
  outpath = NULL,
  ped,
  addEffects,
  domEffects,
  haploMat,
  recombFreqMat,
  ncores = 1,
  ...
)
}
\arguments{
\item{ped}{pedigree data.frame, cols: sireID, damID. sireID and damID must both be in the haploMat.}

\item{addEffects}{vector of _additive_ SNP effects estimate. Ideally vector is named with SNP IDs and matches the order of related SNP matrices.}

\item{domEffects}{vector of _dominance_ SNP effects estimate. Ideally vector is named with SNP IDs and matches the order of related SNP matrices.}

\item{haploMat}{matrix of phased haplotypes, 2 rows per sample, cols = loci, {0,1}, rownames assumed to contain GIDs with a suffix, separated by "_" to distinguish haplotypes}

\item{recombFreqMat}{a square symmetric matrix with values = (1-2*c1), where c1=matrix of expected recomb. frequencies. The choice to do 1-2c1 outside the function was made for computation efficiency; every operation on a big matrix takes time.}

\item{ncores}{If ncores set > 1 parallelizes across families, but beware it is memory intensive and options(future.globals.maxSize=___) may need to be adjusted.}

\item{...}{}
}
\value{

}
\description{
Function to compute predicted add + dom vars for an entire pedigree.
If outprefix and outpath are supplied, writes output to disk so impatient users can see results.
}
