% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{runMultiTraitCrossPredA}
\alias{runMultiTraitCrossPredA}
\title{runMultiTraitCrossPredAD}
\usage{
runMultiTraitCrossPredA(
  outprefix = NULL,
  outpath = NULL,
  CrossesToPredict,
  AddEffectArray,
  traits,
  snpIDs,
  nIter,
  burnIn,
  thin,
  haploMat,
  doseMat,
  recombFreqMat,
  ncores = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
runMultiTraitCrossPredAD
}
