% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{runMultiTraitCrossPredADseqsnps}
\alias{runMultiTraitCrossPredADseqsnps}
\title{runMultiTraitCrossPredADseqsnps}
\usage{
runMultiTraitCrossPredADseqsnps(
  outprefix = NULL,
  outpath = NULL,
  CrossesToPredict,
  AddEffectArray,
  DomEffectArray,
  traits,
  snpIDs,
  nIter,
  burnIn,
  thin,
  haploMat,
  doseMat,
  recombFreqMat,
  ncores = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Function is same as runMultiTraitCrossPredAD but for each family, when prediting variances, only considers segregating SNP.
Makes a HUGE difference in compute speed / mem. usage. Eventually will become runMultiTraitCrossPredAD.
}
