% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{runMultiTraitCrossPredADvpm}
\alias{runMultiTraitCrossPredADvpm}
\title{runMultiTraitCrossPredADvpm}
\usage{
runMultiTraitCrossPredADvpm(
  outprefix = NULL,
  outpath = NULL,
  CrossesToPredict,
  AddEffectArray,
  DomEffectArray,
  traits,
  snpIDs,
  nIter,
  burnIn,
  thin,
  haploMat,
  doseMat,
  recombFreqMat,
  ncores = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
For a set of crosses, with an array of posterior marker effects for multiple traits and a multi-component, namely an additive plus dominance model, predict the genetic means, (co)variances for all traits and all crosses.
For prediction of variances, consider only segregating SNPs to save time and memory.
}
