% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{predOneCrossVarAD}
\alias{predOneCrossVarAD}
\title{predOneCrossVarAD}
\usage{
predOneCrossVarAD(
  Trait1,
  Trait2,
  sireID,
  damID,
  haploMat,
  recombFreqMat,
  predType = "VPM",
  postMeanAddEffects,
  postMeanDomEffects,
  postVarCovarOfAddEffects = NULL,
  postVarCovarOfDomEffects = NULL,
  ...
)
}
\arguments{
\item{Trait1}{string, label for Trait1. When Trait1==Trait2 computes the genomic variance of the trait, when Trait1!=Trait2 computes the genomic covariance between traits.}

\item{Trait2}{string, label for Trait2. When Trait1==Trait2 computes the genomic variance of the trait, when Trait1!=Trait2 computes the genomic covariance between traits.}

\item{sireID}{string, Sire genotype ID. Needs to correspond to renames in haploMat}

\item{damID}{string, Dam genotype ID. Needs to correspond to renames in haploMat}

\item{haploMat}{matrix of phased haplotypes, 2 rows per sample, cols = loci, {0,1}, rownames assumed to contain GIDs with a suffix, separated by "_" to distinguish haplotypes}

\item{recombFreqMat}{a square symmetric matrix with values = (1-2*c1), where c1=matrix of expected recomb. frequencies. The choice to do 1-2c1 outside the function was made for computation efficiency; every operation on a big matrix takes time.}

\item{predType}{string, "VPM" or "PMV". Default = "VPM" for variance of posterior means, this is faster but expected to be less accurate / more biased than the alternative predType=="PMV". PMV requires user to supply a variance-covariance matrix of effects estimates.}

\item{postMeanAddEffects}{list of named vectors (or column matrices) with the posterior mean ADDITIVE marker effects.}

\item{postMeanDomEffects}{list of named vectors (or column matrices) with the posterior mean DOMINANCE marker effects.}

\item{postVarCovarOfAddEffects}{matrix of dimension N SNP x N SNP. ADDITIVE Posterior Sample Variance-Covariance Matrix of Marker Effects Estimates.}

\item{postVarCovarOfDomEffects}{matrix of dimension N SNP x N SNP. DOMINANCE Posterior Sample Variance-Covariance Matrix of Marker Effects Estimates.}

\item{...}{}
}
\value{
tibble with predicted additive and dominance variance for one cross, one variance parameter
}
\description{
Predict the additive and dominance variance or co-variance in one cross.
}
