% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCrossVar.R
\name{runMtCrossVarPredsAD}
\alias{runMtCrossVarPredsAD}
\title{runMtCrossVarPredsAD}
\usage{
runMtCrossVarPredsAD(
  outprefix = NULL,
  outpath = NULL,
  predType,
  CrossesToPredict,
  AddEffectList,
  DomEffectList,
  haploMat,
  recombFreqMat,
  ncores,
  ...
)
}
\arguments{
\item{outprefix}{string, prefix for *.rds file to be written to disk with output. DEFAULT = NULL (no disk write)}

\item{outpath}{string, path to disk location where files should be written. Can be left DEFAULT = NULL (no disk write)}

\item{predType}{string, "VPM" or "PMV". Default = "VPM" for variance of posterior means, this is faster but expected to be less accurate / more biased than the alternative predType=="PMV". PMV requires user to supply a variance-covariance matrix of effects estimates.}

\item{CrossesToPredict}{data.frame or tibble, col/colnames: sireID, damID. sireID and damID must both be in the haploMat.}

\item{AddEffectList}{list of ADDITIVE effect matrices, one matrix per trait, Each element of the list is named with a string identifying the trait and the colnames of each matrix are labelled with snpIDs.}

\item{DomEffectList}{list of DOMINANCE effect matrices, one matrix per trait, Each element of the list is named with a string identifying the trait and the colnames of each matrix are labelled with snpIDs.}

\item{haploMat}{matrix of phased haplotypes, 2 rows per sample, cols = loci, {0,1}, rownames assumed to contain GIDs with a suffix, separated by "_" to distinguish haplotypes}

\item{recombFreqMat}{a square symmetric matrix with values = (1-2*c1), where c1=matrix of expected recomb. frequencies. The choice to do 1-2c1 outside the function was made for computation efficiency; every operation on a big matrix takes time.}

\item{ncores}{If ncores set > 1 parallelizes across families, but beware it is memory intensive and options(future.globals.maxSize=___) may need to be adjusted.}

\item{...}{}
}
\value{

}
\description{
Predict the additive and dominance variances _and_ covariances for a set of crosses, potentially in parallel across families.
Wraps predCrossVarsAD across variance parameters.
The output predicted variances should be suitable to
compute predVar(TGV) = predVar(A) + predVar(D).
}
